update example to clean up layout and usage
