+++ frameworks/nuxt/keycloak-js/components/Links.vue
@@ -1,0 +1,42 @@
+<template>
+  <div class="py-24 sm:py-32">
+    <div class="mx-auto max-w-3xl px-6 lg:px-8">
+      <div
+        class="-mx-6 grid grid-cols-2 gap-0.5 overflow-hidden sm:mx-0 sm:rounded-2xl"
+      >
+        <a
+          href="https://phasetwo.io/docs/introduction"
+          target="_blank"
+          rel="noreferrer"
+        >
+          <div
+            class="p-8 sm:p-10 bg-purple-500/10 text-center text-xl font-semibold text-slate-600/80"
+          >
+            Docs
+          </div>
+        </a>
+        <a href="https://github.com/p2-inc/" target="_blank" rel="noreferrer">
+          <div
+            class="p-8 sm:p-10 bg-purple-500/10 text-center text-xl font-semibold text-slate-600/80"
+          >
+            Github
+          </div>
+        </a>
+        <a href="https://phasetwo.io/blog" target="_blank" rel="noreferrer">
+          <div
+            class="p-8 sm:p-10 bg-purple-500/10 text-center text-xl font-semibold text-slate-600/80"
+          >
+            Blog
+          </div>
+        </a>
+        <a href="mailto:support@phasetwo.io" target="_blank" rel="noreferrer">
+          <div
+            class="p-8 sm:p-10 bg-purple-500/10 text-center text-xl font-semibold text-slate-600/80"
+          >
+            Contact
+          </div>
+        </a>
+      </div>
+    </div>
+  </div>
+</template>
