+++ frameworks/nuxt/keycloak-js/layouts/default.vue
@@ -1,0 +1,11 @@
+<template>
+  <div class="page-bg">
+    <Header />
+    <div class="py-12">
+      <div class="mx-auto max-w-3xl px-6 lg:px-8 text-center">
+        <slot />
+      </div>
+    </div>
+    <Links />
+  </div>
+</template>
