+++ frameworks/nuxt/keycloak-js/pages/index.vue
@@ -1,0 +1,5 @@
+<template>
+  <div>
+    <User />
+  </div>
+</template>
